refactor: isolate `hide battery icon` hook
